\name{CorePE-package}
\alias{CorePE-package}
\alias{CorePE}
\docType{package}
\title{
\packageTitle{CorePE}
}
\description{
\packageDescription{CorePE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CorePE}
\packageIndices{CorePE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CorePE}

Maintainer: \packageMaintainer{CorePE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
